backend/requirements.txt
# Document processing
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
unstructured==0.11.8
python-multipart==0.0.6

# AI and ML
langchain==0.1.16
langchain-community==0.0.32
langchain-openai
sentence-transformers==2.5.1
transformers==4.36.2
torch==2.1.2
huggingface-hub==0.20.3
openai

# Vector database
faiss-cpu==1.7.4
chromadb==0.4.22

# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
nltk==3.8.1
spacy==3.7.2
python-dotenv==1.0.0
pydantic==2.5.0

# Optional: For better text processing
pypdf2==3.0.1